Move component methods to services
